+++ src/routes/index.ts
@@ -1,17 +1,10 @@
 import {Router} from 'express'
-import { signup } from '../endpoints/signup';
-import { login } from '../endpoints/login';
-import { makeFriendship } from '../endpoints/makeFriendship';
-import { undoFriendship } from '../endpoints/undoFriendship';
-import { createPost } from '../endpoints/createPost';
-import { feed } from '../endpoints/feed';
+import { userRoutes } from './UserRoutes';
+import { postRouter } from './PostRouter';
+
 
 export const routes = Router();
 
-routes.post("/signup", signup);
-routes.post("/login", login);
-routes.post("/make-friendship", makeFriendship);
-routes.delete("/undo-friendship", undoFriendship);
-routes.post("/post", createPost);
-routes.get("/feed", feed);
-routes.get("/feed/:type", feed);
+routes.use("/user", userRoutes);
+routes.use("/posts", postRouter);
+
